+++ tests/core/test_config.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from jason.core.config import Config, props
+from jason.core.configuration import Config, props
 
 
 @pytest.fixture